+++ src/App.jsx
@@ -1,11 +1,15 @@
 import './style/App.css';
 import { BrowserRouter } from "react-router-dom";
 import SidebarLayout from './components/Layout';
+import { AuthProvider } from './utils/AuthProvider';
+
 
 function App() {
   return (
     <BrowserRouter>
-      <SidebarLayout />
+      <AuthProvider>
+        <SidebarLayout />
+      </AuthProvider>
     </BrowserRouter>
   );
 }
