+++ src/components/Layout.jsx
@@ -1,38 +1,48 @@
-import { HomeOutlined, AppstoreAddOutlined, AppstoreOutlined, PoweroffOutlined } from '@ant-design/icons';
-import { Layout, Menu, Button } from 'antd';
-import { Route, Routes, Link } from "react-router-dom";
-import Home from './Home';
-import ActiveTickets from './ActiveTickets';
-import NewTickets from './NewTickets';
-import './style/Layout.css';
+import { HomeOutlined, AppstoreAddOutlined, AppstoreOutlined, UserOutlined, UserDeleteOutlined } from '@ant-design/icons';
+import { Layout, Menu } from 'antd';
+import { Route, Routes, Link, useNavigate } from "react-router-dom";
+import { useContext } from "react";
 import { auth } from "../database/authentication";
 import { signOut } from "firebase/auth";
-const { Header, Content, Sider } = Layout;
+import { AuthContext } from "../utils/AuthProvider";
+import Home from '../pages/Home';
+import SignIn from '../pages/SignIn';
+import NewTickets from '../pages/NewTickets';
+import ActiveTickets from '../pages/ActiveTickets';
+import './style/Layout.css';
+const { Content, Sider } = Layout;
 
 
 const SidebarLayout = (props) => {
+  const {currentUser} = useContext(AuthContext);
+  let navigate = useNavigate()
+
   const logout = () =>{
     signOut(auth)
-    .then(alert('Cerró su sesión correctamente'))
+    .then(() =>{
+      alert('Cerró su sesión correctamente')
+      setInterval(navigate('/'), 1000)
+    })
   }
-  const menuItems = [
+  const loggedOutMenuItems = [
     {label: <Link to='/'>Página principal</Link>, key: '1', icon: <HomeOutlined />},
-    {label: <Link to='/nuevos-tickets'>Nuevos Tickets</Link>, key: '2', icon: <AppstoreAddOutlined />, },
-    {label: <Link to='/tickets-activos'>Tickets Activos</Link>, key: '3', icon: <AppstoreOutlined />, }]
-  
+    {label: <Link to='/ingresa'>Ingresa</Link>, key: '2', icon: <UserOutlined />},] 
+  const loggedInMenuItems =[
+      {label: <Link to='/nuevos-tickets'>Nuevos Tickets</Link>, key: '3', icon: <AppstoreAddOutlined />, },
+      {label: <Link to='/tickets-activos'>Tickets Activos</Link>, key: '4', icon: <AppstoreOutlined />, },
+      {label: 'Cerrar sesión', key: '5', icon: <UserDeleteOutlined />, onClick: logout, danger: true}]
     return (
     <Layout style={{ minHeight: '100vh' }}>
       <Sider collapsed={true}>
         <div className="logo" />
-        <Button onClick={logout} icon={<PoweroffOutlined />}>Logout</Button>
-        <Menu theme="dark" mode="inline" items={menuItems} />
+        <Menu theme="dark" mode="inline" items={currentUser ? loggedInMenuItems : loggedOutMenuItems}/>
       </Sider>
       <Layout className="site-layout">
-        <Header className="site-layout-background" style={{padding: 0,}} />
-        <Content style={{margin: '0 16px',}}>
+        <Content style={{margin: '16px 16px'}}>
           <div className="site-layout-background" style={{padding: 24,minHeight: 360,}}>
             <Routes>
               <Route path='/' element={<Home />} />
+              <Route path='/ingresa' element={<SignIn />} />
               <Route path='/tickets-activos' element={<ActiveTickets parentProps={props}/>} />
               <Route path='/nuevos-tickets'  element={<NewTickets parentProps={props}/>} />
             </Routes>
