+++ src/index.js
@@ -1,4 +1,4 @@
-import React, {StrictMode} from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './style/index.css';
 import App from './App';
@@ -7,11 +7,9 @@ import { Provider } from 'react-redux'
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
-      <StrictMode>
-            <Provider store={store}>
-                  <App />
-            </Provider>
-      </StrictMode>
+      <Provider store={store}>
+            <App />
+      </Provider>
 );
 
 // If you want to start measuring performance in your app, pass a function
