+++ src/pages/Home.jsx
@@ -1,6 +1,4 @@
-import { Card, Row, Col, Divider } from 'antd';
-import SignIn from "./SignIn";
-
+import { Card, Row, Col } from 'antd';
 const { Meta } = Card;
 
 
@@ -22,10 +20,6 @@ export default function Home(){
     <div>
       <Row justify='center'>
         {Cards}
-      </Row>
-      <Divider orientation='center'>Ingrese a su cuenta</Divider>
-      <Row justify='center' style={{margin: '20px 0'}}>
-        <SignIn />
       </Row>
     </div>
   )
