+++ src/utils/AuthProvider.js
@@ -1,0 +1,26 @@
+import React, { useEffect, useState } from "react";
+import { onAuthStateChanged } from "firebase/auth";
+import { auth } from "../database/authentication";
+
+export const AuthContext = React.createContext();
+
+export const AuthProvider = ({children}) => {
+    const [currentUser, setCurrentUser] = useState(null);
+
+    useEffect(() => {
+      onAuthStateChanged(auth, (user) => {
+        if (user) {
+          setCurrentUser(user);
+        } else {
+          setCurrentUser(null);
+        }
+      });
+    console.log(currentUser);
+    }, [currentUser]);
+
+    return (
+    <AuthContext.Provider value={{ currentUser }}>
+      {children}
+    </AuthContext.Provider>
+  );
+};
